+++ MordoEngine/src/Scene/Editor/Controllers/EditorCameraController.h
@@ -1,0 +1,14 @@
+#pragma once
+#include "../../../Input/Input.h"
+#include "../../../Camera/Camera.h"
+#include <memory>
+
+class EditorCameraController
+{
+private:
+	std::weak_ptr<Camera> m_Camera;
+
+public:
+	EditorCameraController(std::weak_ptr<Camera> camera);
+	void Update(float deltaTime , float velocity);
+};