+++ MordoEngine/src/Scene/Editor/EditorScene.cpp
@@ -4,10 +4,10 @@
 #include "../../Input/Input.h"
 
 EditorScene::EditorScene(std::shared_ptr<terrain::Terrain> terrain, std::shared_ptr<Camera> camera, std::shared_ptr<Renderer> renderer)
-	: Scene("terrain"), m_Terrain(terrain), m_Camera(camera), m_Renderer(renderer),
-	  m_TerrainSelector(std::make_unique<AreaSelectorRenderer>(camera))
+	: Scene("terrain"), m_Terrain(terrain), m_Camera(camera), m_Renderer(renderer)
 {
-
+	m_TerrainSelector = std::make_unique<AreaSelectorRenderer>(m_Camera);
+	m_CameraController = std::make_unique<EditorCameraController>(m_Camera);
 }
 
 
@@ -30,27 +30,8 @@ void EditorScene::Render()
 
 void EditorScene::Update(float deltaTime)
 {
-	glm::vec3 newCameraPosition = m_Camera->GetPosition();
 	float velocity = 100.0f * m_Terrain->GetWorldScale() * deltaTime;
-
-	if (Input::KeyDown(GLFW_KEY_W)) {
-		newCameraPosition += m_Camera->GetForward() * velocity;
-	}
-	if (Input::KeyDown(GLFW_KEY_A)) {
-		newCameraPosition -= m_Camera->GetRight() * velocity;
-	}
-	if (Input::KeyDown(GLFW_KEY_S)) {
-		newCameraPosition -= m_Camera->GetForward() * velocity;
-	}
-	if (Input::KeyDown(GLFW_KEY_D)) {
-		newCameraPosition += m_Camera->GetRight() * velocity;
-	}
-
-	if (newCameraPosition != m_Camera->GetPosition()) {
-		m_Camera->SetPosition(newCameraPosition);
-	}
-	m_Camera->ProcessMouseMovement(Input::MouseDeltaX(), Input::MouseDeltaY());
-
+	m_CameraController->Update(deltaTime, velocity);
 	m_TerrainSelector->Update(*m_Terrain);
 }