+++ MordoEngine/src/Scene/Game/GameScene.cpp
@@ -6,31 +6,14 @@
 GameScene::GameScene(std::shared_ptr<terrain::Terrain> terrain, std::shared_ptr<Camera> camera, std::shared_ptr<Renderer> renderer) 
 	: Scene("terrain"), m_Terrain(terrain) , m_Camera(camera) , m_Renderer(renderer) , m_Sun(std::make_unique<Sun>(0.9f))
 {
-
+	m_CameraController = std::make_unique<EditorCameraController>(m_Camera);
 }
 
 void GameScene::Update(float deltaTime)
 {
 	m_Sun->Update(deltaTime);
-	glm::vec3 newCameraPosition = m_Camera->GetPosition();
 	float velocity = 100.0f * m_Terrain->GetWorldScale() * deltaTime;
-
-	if (Input::KeyDown(GLFW_KEY_W)) {
-		newCameraPosition += m_Camera->GetForward() * velocity;
-	}
-	if (Input::KeyDown(GLFW_KEY_A)) {
-		newCameraPosition -= m_Camera->GetRight() * velocity;
-	}
-	if (Input::KeyDown(GLFW_KEY_S)) {
-		newCameraPosition -= m_Camera->GetForward() * velocity;
-	}
-	if (Input::KeyDown(GLFW_KEY_D)) {
-		newCameraPosition += m_Camera->GetRight() * velocity;
-	}
-
-	if (newCameraPosition != m_Camera->GetPosition()) {
-		m_Camera->SetPosition(newCameraPosition);
-	}
+	m_CameraController->Update(deltaTime, velocity);
 	m_Camera->ProcessMouseMovement(Input::MouseDeltaX(), Input::MouseDeltaY());
 }